requirements.txt
asgiref==3.8.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
Django==4.1.13
django-extensions==3.2.3
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
duplicity==0.8.21
fasteners==0.14.1
fonttools==4.53.1
future==0.18.2
graphviz==0.20.3
html5lib==1.1
idna==3.3
importlib-metadata==4.6.4
joblib==1.4.2
kiwisolver==1.4.5
language-selector==0.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.9.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
olefile==0.46
packaging==24.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==3.12.3
PyNaCl==1.5.0
pyparsing==3.1.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
scikit-learn==1.5.1
scipy==1.14.0
scour==0.38.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.2.4
ssh-import-id==5.11
threadpoolctl==3.5.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
urllib3==1.26.5
usb-creator==0.3.7
webencodings==0.5.1
websockets==12.0
xdg==5
xkit==0.0.0
zipp==1.0.0
